Updates to Treaties PR
This responds to @flooie's comments on the last version of the PR. I pulled out the (?:No\\.|Number) group and replaced it with $number_label, which I added to regexes.json. I also added some new examples for treaties using Number instead of No. Finally, I cleaned up some law citations that were not to treaties but which tried to match No. without escaping the period. I replaced them with $number_label.